Quick responsive fix
The main page will now reduce to 2 columns on tablets and 1 column on
phones. The site's not fully responsive yet, but it's a quick and
convenient fix.

The <li> elements are styled as inline-block, so they can't have
whitespace in between, otherwise the 50% width with a space would exceed
the width of the parent.